# [perf] `msmReduceChunk` in Jac instead of JacExt