Currently we implement `msmReduceChunk` in extended Jacobian coordinates. It costs `(s-1)c` doublings, `(s-1)` additions and `1` conversion from JacExt to Jac (`2`sq + `2`mul), where `s` is the chunk size and `c` the window size. For large MSM instances, it is better to do it in Jacobians or at least the doublings in Jacobians, as they cost `3` mul less. The addition, however, costs `2` mul more. This PR trades off `10(s-1)c+14(s-1)+4` muls for `7(s-1)c+16(s-1)+4s` muls.