@yelhousni 2022-11-04
Closing this because the benchmarks are not as good on `x86` as on `M1`. e.g. BN254 G1-MSM:
```
M1:
benchmark                          old ns/op      new ns/op      delta
BenchmarkMultiExpG1Reference-8     1336971708     1235571875     -7.58%
```
and
```
AWS z1d.large:
benchmark                          old ns/op      new ns/op      delta
BenchmarkMultiExpG1Reference-2     4399459569     4361069253     -0.87%
```